+++ src/components/WorkflowEditor/index.tsx
@@ -247,7 +247,7 @@ const WorkflowEditor = (): JSX.Element => {
                 )
             ) : (
                 <>
-                    <Box sx={{ padding: "20px", height: "8vh", borderBottom: "solid 1px black" }}>
+                    <Box sx={{ padding: "20px", height: "4vh", borderBottom: "solid 1px black" }}>
                         <Stack direction="row" sx={{ justifyContent: "space-between" }}>
                             <Box sx={{ alignContent: "center" }}>
                                 <Stack direction="row" spacing={2}>