+++ src/routes.tsx
@@ -4,7 +4,8 @@ import Loading from "@app/components/Loading";
 const LazyHome = lazy(() => import("@app/components/Home"));
 const LazyWorkflowEditor = lazy(() => import("./components/WorkflowEditor"));
 const LazyProjectDatasetPage = lazy(() => import("@app/components/Project/Resource/DatasetPage"));
-const LazyProjectPage = lazy(() => import("@app/components/Project/ProjectPage"));
+// const LazyProjectPage = lazy(() => import("@app/components/Project/ProjectPage"));
+const LazyProjectDashboard = lazy(() => import("@app/components/Project/ProjectDashboard"));
 const LazyProjectHazardsPage = lazy(() => import("@app/components/Project/Resource/HazardPage"));
 const LazyProjectVisualizationsPage = lazy(() => import("@app/components/Project/Resource/VisualizationPage"));
 const LazyProjectWorkflowsPage = lazy(() => import("@app/components/Project/Resource/WorkflowPage"));
@@ -31,7 +32,8 @@ const routes: { [key: string]: import("react-router-dom").RouteProps } = {
     "/project/:id": {
         element: (
             <Suspense fallback={<Loading />}>
-                <LazyProjectPage />
+                {/* <LazyProjectPage /> */}
+                <LazyProjectDashboard />
             </Suspense>
         )
     },